Estilização adicional de telas de listagem e de detalhes, organização de componentes, revisão do codigo, documentação de funcionalidades e descrição do projeto

+++ app/src/main/java/com/example/pokedex/data/remote/dto/PokemonDetailDto.kt
@@ -4,8 +4,8 @@ data class PokemonDetailDto(
     val id: Int,
     val name: String,
     val sprites: SpritesDto,
-    val weight: Int,
-    val height: Int,
+    val weight: Float,
+    val height: Float,
     val stats: List<StatDto>,
     val types: List<TypeSlotDto>
 ) {
